Add missing 'update_tasks' to the OVN dbs pacemaker service
Presently the OVN dbs pacemaker service file doesn't include the update_tasks
section and it neither tags the docker image to be used with pcmklatest.
Because of which when a minor update is run, the latest docker image for
ovn-dbs-bundle is never pulled.

This patch addresses these issue. It does 3 things

1. It adds update_tasks section similar to other pacemaker bundle services.
It adds additional few steps in step 5 (which is not the case with other
pacemaker bundle services) to update the ovn bundle service to use the
image tagged with pcmklatest if it not yet using. These additional steps
are required for OVN deployments deployed before this patch and wants to
do a minor stack update.

2. It adds the required tasks in "upgrade_tasks" section similar to
other pacemaker bundle services.

3. Adds docker_config step to tag the image with pcmklatest tag.

Change-Id: Idbda3bedff57376b74269ab3470d2324b804ffd4
Closes-bug: #1775686
(cherry picked from commit c90e0a7197e5f84d4f0cf1e284866eba48c093d4)