Set Ceph pools rule_name to replicated_rule
By setting the value of rule_name explicitly, we prevent backward
incompatibility issues because the default which ceph-ansible uses
might fit a particular version of Ceph, not all.

Note that the actual value for rule_name is different in the Pike
backport, deploying Ceph Jewel, because it needs to be set to
"replicated_ruleset" and not "replicated_rule" as with Ceph Luminous.

Change-Id: I275c1ca53ea79eea607cbbb58aa21cae6d6be80b
Closes-Bug: 1776252
(cherry picked from commit dcf558628f62f12829044dbeee56047f485585f0)