Parallelize server pre and post steps
Update the play for the server pre and post steps so that the tasks run
in parallel across all roles, instead of doing one role at a time. By
not using the "when" attribute, and relying on the tripleo_role_name var
for the list of deployments, we can force these tasks to run in parallel
across all roles.

Change-Id: I83a4deaa68d5788edb5ab13652bb30c762f337d8