Set ODL API restart policy to unless-stopped
On minor update, when docker RPM or configuration gets updated, docker
daemon restarts, and if container has restart policy `always`, it is
started automatically. This meant that our previous aim of stopping
the container and cleaning its cache was effectively undone by the
docker daemon restart later during execution of update_tasks.

Setting the restart policy to `unless-stopped` keeps the container
stopped on docker daemon restarts if the container was stopped
intentionally.

Co-Authored-By: Janki Chhatbar <jchhatba@redhat.com>
Co-Authored-By: Tim Rozet <trozet@redhat.com>
Change-Id: Id033cbeef46bded13f4fd6c06eb3a244634f0e7b
Partial-Bug: #1775919