Remove unuseful become: true from deploy-steps
The ansible command generated in ansible-playbook-command.sh by default
have "--become" in it.
This commit removes "become: true" where is used to avoid confusion in
deploy steps. Today we explicitly set "become: false" in deploy-steps.j2
for certain actions, so there's no meaning of having also "become: true"
for the other ones.
We have a release note [1] that explains why the "become" was
introduces, but maybe we can revisit it.

[1] releasenotes/notes/use-become-true-in-deploy-steps-playbook-01decb18d895879f.yaml

Change-Id: Ic666b4ecaecf0591dd8bb0406f239649b20b9623