Reduce verbosity during mysql bootstrap, non-HA
During the initial deployment, a one-time container is used to bootstrap the
mysql databse on disk, create the required users and set their password. The
script that runs that is too verbose and logs some credentials in the
container's logs and in the journal.

Use kolla_extend_start directly instead of kolla_start to stop tracing shell
commands and reduce logging to the bare minimum for troubleshooting.

This is the non-HA counterpart of I90827feff0d1b9fd8badb72e68e4c8dd8db8aea5,
which is now also used to set up containerized undercloud.

Change-Id: Ib583fa92d279fc3166cceaf58e152b5cce315d44
Closes-Bug: #1765339