Log memcached to stdout/err only
Memcached logging configuration is not friendly.
We cannot use configuration options as they seemed to be
ignored because of a bug. We cannot use redirection >>
as this breaks logrotate signalling SIGHUP to processes
and expecting them to handle the open file descriptors
associated with truncated/removed logfiles - this does
not work with redirected output into a file.

The only option is not log memcached into a file. There is
not much of actual value in the log. Keep it logging to
stdout/stderr in its container, which will get captured in
the docker logs (and journald via the logging backend).

Co-authored-by: Sergii Golovatiuk <sgolovat@redhat.com>
Co-authored-by: Alex Schultz <aschultz@redhat.com>
Closes-bug: #1776889

Change-Id: I3b5aec6b7f7c5fa4f84e3c3eb243c155a6189e2c
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>