Activate memcached debug only when using MemcachedDebug param.
Do not activate it with Debug anymore.  We are now pushing memcached
log into stdout where they will be collected by journald.  When
activating debug, it will log there all the requests/replies to/from
memcached.  This will make the journal grows fast and clutter it
will (most of the time) useless information.

So if memcached debug is needed for some reason let the operator says
so explicitly.

Change-Id: I85b6ca421d326c6e764fade66726407b905c796d