Make BIND /var dir persistent
BIND stores dynamically created zones in /var/named by default,
which means that this directory needs to be persistent across
container restarts.  However, we can't just bind mount /var/named
into the container because /var/named is empty on the host, but in
the container the BIND package populates some necessary initial files
in it.

To address this, we can point BIND at /var/named-persistent and
copy all of the initial config from /var/named at container start,
just like we do for other configs.

Change-Id: Ic17fb812468f2eb66d9c348b51dfa73a1f216a70