Disable pid file usage in the swift_rsync container
The pidfile is useless within containers, but prevents a restart if the
process has been stopped unclean. This happens for example when a
restart takes longer than the default timeout of 10 seconds.

This is mainly a workaround because the pid file setting is hardcoded in
the used Puppet module template. It should be removed later once the
setting can be disabled cleanly with an updated Puppet rsync module.

Related-Bug: 1724559
Change-Id: Iecf855a785bc1f787234e4e54430c929cb9cb906