Fix incorrect Redis VIP declaration in fixed-ip-vips-v6.yaml
This patch fixes the incorrect Redis VIP declaration in
fixed-ip-vips-v6.yaml. The file incorrectly has vip.yaml
for the RedisVipPort instead of vip_v6.yaml, although the
fixed IP is an IPv6 address. This causes an error, and
conflicts with network-isolation-v6.yaml, so the deployment
works if network-isolation-v6.yaml is included after the
fixed-ip-vips-v6.yaml, but fails if it is included before.

Change-Id: Iffabbedfda77bc5cc4793700c459fb51b14f62d9
Closes-bug: 1778120