Upload amphora image in RAW format if RBD backend
When deploying an overcloud with Octavia enabled, the amphora image is
uploaded to Glance as QCOW2. When using the RBD backend for Glance and
Nova, this causes the image to be converted to RAW on each deployment.
With slower storage, this leads to load balancer creation to fail with a
compute timeout exception from Octavia.

This patch makes TripleO oppiniated. If Ceph is being used, it will
automatically convert the QCOW2 file to RAW before uploading to Glance.
The upfront cost of conversion to RAW and uploading to Glance is worth
the performance improvements.

Closes-Bug: #1778303
Depends-On: https://review.openstack.org/#/c/577573/
Change-Id: If36c87e783b532b1e8b368d8e18f5a83b581b933