Execute post_update_tasks in update playbook
We drop the post_update_steps_playbook, and execute post_update_tasks
as part of the update_steps_playbook. This will ensure that
post_update_tasks are executed, and they're executed in accordance
with the `serial: 1` ordering that update_steps_playbook is using. (We
want to avoid an issue similar to what we've had in bug #1776206).

Change-Id: I15a984172cd5532bc966269d8c68f27b5703733e
Closes-Bug: #1778471