Add the ability to scaleup the openshift stack
Refactored the inventory generation to allow for scaling up the
openshift stack. They deployment process now checks whether the
origin-node container is already running in the openshift nodes. All
the nodes where origin-node is not present are considered as new.

The new_masters/new_nodes sections will be populated accordingly and
the `scaleup.yml` playbook will be used instead of the `deploy_cluster`
one *only* if the stack is being updated.

Unfortunately, openshift-ansible's implementation is not as idempotent
as TripleO's, which requires differentiating between deployments,
scaleup and between new and old nodes.

Change-Id: I15c4e855a3ccfba6ce8d516b5b59e7508c4885cc