Clean up env files for Cinder's Netapp backend
Remove the misleading comment from the cinder-netapp-config.yaml file in
the environments directory that claimed the file is deprecated in favor
of the 'sample-env-generator' file in the environments/storage directory.
The sample-env-generator has not been widely adopted, and the Netapp
file in the environments directory is still being maintained.

Update the sample-env-generator files so they specify the
OS::TripleO::Services::CinderBackendNetApp composable service.

Closes-Bug: 1779397
Change-Id: If0343e96700bf13c0d49c3ae3c311ada1c26a0c0