OVN: Remove environment files which deploy OVN dbs in non HA
In a 3 node controller deployment OVN db servers are started only on one node
and if that node goes down, we hit the down time for ovsdb-servers and neutron-server.

So it is highly recommended to deploy OVN db servers as a pacemaker HA resource.

Conflicts:
	environments/neutron-ml2-ovn-dvr.yaml
	environments/neutron-ml2-ovn.yaml
	environments/services-docker/neutron-ovn-dvr.yaml
	environments/services-docker/neutron-ovn.yaml
	environments/services/neutron-ovn-dvr.yaml
	environments/services/neutron-ovn.yaml

Change-Id: Ia2f50d539bf6afbeb67610a925f9f4ea0849d735
Closes-bug: #1779123
(cherry picked from 819805d708cccc6e1419a47b16fc012955ce9156)