Add support for {{role.name}}IPs in Composable Role
Use the {{role.name}}IPs for network port resources in
Composable Role to enable setting predictable ip address
on network ports.

We currently support predictable ip addressing using the
from_pool templates. The from_pool creates a 'fake' neutron
port. Because of this we loose access to properties of
the ports subnet, properties that is useful to simplyfy
advanced configurations such as routed spine-and-leaf.

Creating a neutron port also open the possibility to use ML2
plugins in the future (Neutron Ansible ML2) to do port
binding that configures network devices.

The parameter (interface) to use this is the same as with
from_pool. Example:

parameter_defaults:
  {role.name}IPs:
    {network.name_lower}:
    - 10.0.0.1
    {network.name_lower}:
    - 172.16.1.1

We just don't need to override the resource registry. I.e
the following is not used/required.
resource_registry:
  ``OS::TripleO::{{role.name}}::Ports::{{network.name}}Port``

NOTE: This does not remove the from_pool templates. We may
      want to deprecate them, but in that case we need to
      ensure it is possible to upgrade an environment that
      used them by changeing the resource registry to use
      the default port templates in this change.

Related: blueprint tripleo-routed-networks-templates
Change-Id: I3f9b55ffbf6b6626b4d0dfbec3867811cb74de14