Do not disable ipv6 on loopback interface for epmd
Currently, as discussed in [1], the Erlang Port Mapper Daemon requires that
the loopback interface supports IPv6 in order to initialize properly.
Without that, rabbitmq-server cannot start and deployment fails at step 2.

Until the startup behaviour of epmd is amended, do not disable inet6 support
on loopback device to workaround the problem.

Closes-Bug: #1780065

[1] https://bugs.launchpad.net/ubuntu/+source/rabbitmq-server/+bug/1312507/comments/15

Change-Id: I200acaeaca15e47dc2e3b2462a6254e85477ffb0