host_routes using get_attr (Composable Networks)
Uses get_attr on the port resource to resolve attribute
value from neutron subnet 'host_routes' attribute and
pass it to the parameter '{{network.name}}InterfaceRoutes'
in network configuration templates.

A conditional is used in puppet/role.role.j2.yaml. The
user provided parameter value is used whenever it is
not the default: []. This allow advanced user's to
override the routes in neutron.

Co-Authored-By: Dan Sneddon <dsneddon@redhat.com>
Partial: blueprint tripleo-routed-networks-templates
Change-Id: Ie44b211c4aeab9ca79d7994f31961e34aa3517e6