Add AllNodesExtraMapData parameter
The AllNodesExtraMapData parameter is used to inject additional
hieradata into the all_nodes hierdata file on each node. The injected
data will be deeploy merged with the calculated all_nodes data for the
stack.

The parameter can be taken advantage of for split-controlplane use cases
where the hieradata from the control stack needs to be populated into
the separate compute stacks.

To easily get the hieradata out of the control stack, a new stack output
is added, AllNodesConfig.

Partially Implements: blueprint split-controlplane

Change-Id: I7b865bf82520006eef3ac2f36df34b1f3c34e642