Cleanup hosts name on initial install
When we initially install our heat hosts configuration, we need to make
sure the short hostname for the host is not currently defined within the
hosts file as it will conflict with our definitions later in the file.

We only do this on initial installation as we don't want to touch the
file after the initial deployment outside of the matching host entries.
This also is restricted to just the short hostname which is consumed via
the configuration management for determining bootstrap configurations as
well as used by some services during startup (rabbitmq)

Change-Id: If55cf8f90ee7be4acd40fda1f72bb1f31d218b57
Related-Blueprint: all-in-one