Fluentd:Read log files from non containerized services
This only happens on Pike as It's the only release where
containerized and non-containerized services are deployed
mixed.

Fluentd is not configured properly as it expects to have
all the services containerized. This patch added with
the puppet one adds an extra transformation to be able
to set fluentd properly.

Change-Id: Ic33a397f68d6b429b69b1849f9998672b6e97b0a
Related-Bug: #1742651
Depends-On: I3e104fbaba0b7944deccf39fd87f3efaf1e21649