Enable deep_compare of pcmk resources by default
By settings these hiera keys we will actually do a deep comparison
of all pacemaker resources and update them accordingly.

This means that rerunning a deploy against an existing overcloud
will actually detect if a pacemaker resource needs updating and
will update the definition and restart of said resource. The code
will wait for the resource to be up (aka the cluster to settle
after an update).

Depends-On: I62d5662327333e17e1c32a029695b3d3a904ca10
Change-Id: I0b9ed429e6b1bd4a62b4462bf6312782f29b7d6b