Iterate once in network-environment.yaml
This replaced multiple for loops with a single loop.
This changes the ordering in the rendered file to have
one section per network, instead of one section per
perameter such as VLAN, AllocationPools.

It also reduces the use of jinja conditionals. This
results in some duplicated comment strings in the
jinja template.

This is a micro optimization, but this makes sense
as a follow on will add the possibility to add
multiple subnets on each network in network_data.yaml.
I.e we avoid additional nested loops over the subnets.

Related: blueprint tripleo-routed-networks-templates
Change-Id: I0f11acaf7c25ff5d240db1daae2dc2b051c0a3b2