Add secondary DNS server to disable-unbound environment
The Cloudflare DNS we are using in the disable-unbound CI
environment for OpenShift works fine in CI, but my ISP seems to
block it. This makes reproducing that job locally difficult. I had
success with the secondary DNS from Cloudflare, so this patch just
adds that to the resolv.conf.

If we were going to keep this disable-unbound solution for a long
time it would probably be better to have this be a template and
allow user configuration. However, my understanding is that this
is a temporary solution, so investing in complicated patches to
wire in a configuration option that will go away seems like
wasted effort.

Change-Id: I7b93efcd76b651807dff3c18885b8d291feffd3e