L3 routed networks - network.j2 (2/3)
Add parameters for the subnets of each network in network_data
yaml. Associates the base subnet of each network with the first
segment of the network. (All networks have an implicit first
network segment when they are created.)

Additional segments and subnets defined in the ``subnets`` key
of each composable network is created with subnet to segment
association.

Partial: blueprint tripleo-routed-networks-templates
Change-Id: I53559ed1445f66fa50508ac9898cbec07914b3fc