Fix HostnameMap lookup - replace str_replace with yaql
Using str_replace when looking up hostnames in
HostnameMap may return unpredictable results.
For example if the map contain:

HostnameMap:
  overcloud-compute-12: cmp-12-london
  overcloud-compute-123: cmp-123-berlin

When we do string replace for a hostname that is
not in the map (example: 'overcloud-compute-124')
it will match on the substring. The result would
be 'cmp-12-london4'.

This change replaces the str_replace with a yaql
expression that will return the value of they key
matching the hostname provided. If the key is not
in the map, the hostname is returned.

Closes-Bug: #1781560
Change-Id: Ia3a6a3539cd566131cd81b4d755190aa5b61a573