mysql/undercloud: add condition for upgrade
The MySQL commands that we run during the upgrade of MySQL are useful
when upgrading a non-containerized undercloud (from instack-undercloud)
to a containerized undercloud, because we need to rename the Neutron
database to match with the defaults in THT.
So these 2 steps are only useful when MariaDB was actually running on
the host and not in a container.
Adding a conditional that checks if MariaDB is running and then we can
run these MySQL commands will ensure we can run the undercloud upgrade
in Rocky and beyond.

Change-Id: I34f3d5bae4838793fcbbc8832f288b11e6cd81f0