Update/upgrade prepare/converge in config-download era
We inherit the `prepare` commands from deploy, in Rocky we will
automatically get config-download env file there so no need to
explicitly disable DeploymentSteps in *-prepare.yaml files too. We
should also disable the WorkflowSteps in config-download environment,
as they too aren't meant to be used with config-download.

Similarly, for `converge` commands in the *-converge.yaml files we
shouldn't re-enable the DeploymentSteps and WorkflowSteps, as that
would go against using config-download.

Change-Id: I89f54ba9a0a15afd26da4d4b26aba8ef09ebf04b