Ensure Cinder has a default volume type
This patch adds a CinderDefaultVolumeType parameter, and does two
things:

1. Configures the default_volume_type in /etc/cinder/cinder.conf
2. Ensures the volume type exists

The cinder::api puppet module supports configuring the default_volume_type,
and this patch simply sets the corresponding hiera value.

The cinder::type puppet resource is used to ensure the default volume
type exists. This is done using docker_puppet_tasks so that it executes
after the Cinder services are running. This is necessary because the
cinder::type puppet provider executes osc commands.

Closes-Bug: #1782217
Depends-On: Ia23996abefdd1410fb86f04ed84a314f4364339c
Change-Id: I99496aa82bb66eccd86f0e482f1fbdc656c07c5d