Fix containerized ironic bind-mounts
We obvserve errors "DriverLoadError: Driver, hardware type or interface
ilo-pxe could not be loaded" as it uses the host level access to
/var/lib/ironic, which is owned by root.  When we have the
host/container(s) shared access to a bind mounted dir, it needs to be
shared instead of the default private mode.

Additionally, fix the missing /var/lib/ironic ownership change in
the ironix-pxe's kolla config.

Related-bug: #1782267

Change-Id: I5ad4956174346f024945852da09f3841c3378c2b
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>