Split designate envs
Because the designate parameters will always need to be edited for
a deployment, a copy of the environment must be made.  However,
because there were resource_registry entries in the previous
enable-designate environments those relative paths would become
invalid if the file was moved.  Splitting the resource_registry
entries from the user-configured parameters should eliminate this
problem.

Change-Id: I8817a36e20e7a75b340a0d6cb0abf09e57b1fd63