copy ceph config in manila-share container bundle
The pacemaker manila-share docker configuration was
set up on the assumption that /etc/ceph on the host
is bind-mounted directly to /etc/ceph in the manila-share
container, but since [1] this is no longer true.  Ceph
configuration follows the pattern wherein the ceph
configuration and keys are bind-mounted into the container
at /var/lib/kolla/config_files/src-ceph/.

So set the docker kolla config for the pacemaker
bundled manila share to copy the contents of
/var/lib/kolla/config_files/src-ceph to /etc/ceph in
the container as is done in the non-pacemaker service
and as is done with cinder-volume service with both
the pacemaker and non-pacemaker versions of the
service.

Closes-Bug: #1783149

[1] I66a76d3226d5a1d3e0b5387326305d1f3313a340

Change-Id: If6cb5e7bd0a5d7952ab43dffddb0e8eaf4952a0d
(cherry picked from commit ba5a57831a43477afc4edecbaf3d33638830496e)