Set default number of rabbitmq queues to CEIL(N/2)
The default used to be -1 which translated to 'ha-mode: all'.
That means that every queue is mirrored on all controllers.

By setting this to zero we are moving to 'ha-mode: exactly' with
the number of queues to be set to CEIL(N/2):
$nr_ha_queues = CEIL(nr_of_rabbit_nodes/2)
set_policy='ha-all ^(?!amq\\.).* {\"ha-mode\":\"exactly\",\"ha-params\":${nr_ha_queues}}'

This way the queues will not be copied around all servers,
but just to a subset of them. This still provides the necessary
resilience in case of a controller crash, but is less demanding
in terms of performance (and likely triggers fewer bugs in rabbit)

Co-Authored-By: John Eckersberg <jeckersb@redhat.com>

Change-Id: If04a7bf24a5c94ec15726843a8a4d2504b656cb7