undercloud: revert to using the iscsi deploy interface by default
The direct deploy interface looks promising in the scale tests so far,
but it prevent local testing and PoC with nodes with less than 8 GiB
RAM because it has to convert the overcloud-full image in memory.

This change changes back to the iscsi deploy interface, leaving
the direct deploy interface fully configured and opt-in.

This patch will likely be reverted in Stein.

Change-Id: I5f8126474ab15a310b4ba305c4d537b93e9f0399
Related-Blueprint: ironic-direct-deploy