Remove HostnameFormatDefault from NFV roles
Value of HostnameFormatDefault defined in role/*.yaml files is
redundant as the default value in the templates sets the same
value as the role name in the lowercase. With the patch
Ifa60eae1ad09b2ceac207114c40c714a6fc67cbc merged, it is possible
to generate multiple roles out of the defined roles. The default
value of the {{role.name}}HostnameFormat in overcloud.j2.yaml will
apply the same format with role name in lower case, having this
default will override all generated roles with same format.
Removing the default from roles will ensure that the default value
is set as per the role name.

As NFV roles are mostly hardware associated, it is important change
for these roles, though the same can be extended to all role
definitions.

Change-Id: I701bc86034a3b75ca05ae08214dcbb2d9c6c7483