Avoid ifcfg files modification when a reboot is not required
A reboot is required after updating the kernel args. In order for
a successful reboot, ifcfg files of unused interfaces are update
so that it reboot will not fail in network service start. These
ifcfg file modifications are not required during the second run
when the reboot is not required (after kernel args set and reboot).
Added a check to avoid the file modification when it is not required.

Change-Id: Ib2381c434d2e14c99f30beb7963acbbe58c4cb64